requirements.txt
# Backend Dependencies
django>=5.0.0
djangorestframework>=3.14.0
pandas>=2.1.0
numpy>=1.24.0
reportlab>=4.0.0
django-cors-headers>=4.3.0

# Desktop Client Dependencies
requests>=2.31.0
PyQt5>=5.15.0
matplotlib>=3.7.0
